Prevent infinite loop in puglProcessEvents() (OSX)
(Prerequisite to use keyboard input on OSX is to embed the program to an App bundle.)

Problem:
Using the example program pugl_test, 'q' can not quit the program.
Hitting (X) on the window closes the window but the process is still running and
has to be killed manually.

To track this down, the following "logging" was introduced to pugl_test.c:
(not part of this patch!):

//main loop
while (quit==0) {
fprintf(file,".");fflush(file);
	puglWaitForEvent(view);
fprintf(file,"-");fflush(file);
	puglProcessEvents(view);
fprintf(file,"*");fflush(file);
}

It showed that the loop never turns, and quit can never be evaluated, also
the view will not be destroyed.

It seemed that puglProcessEvents() is blocking.

Removing the while(true) in puglProcessEvents() shows expected behaviour.
Logs show that now the while(quit==0) loop turns and gets all events when
the application has focus. It makes it possible to cleanly quit the program.