Fix tracking area implementation on MacOS
I am not sure why updateTrackingAreas was being spammed before, but that caused
the event loop to constantly tick when the mouse was hovered over the window
despite nothing else happening.

Everything seems to work fine without it, and this seems to match other code,
so remove it.  Also call super in updateTrackingAreas as suggested by the
documentation.