Add PuglWorld
The old API was broken for programs that manage multiple views, since it was
impossible to wait for events on any view.  There are also several functions in
the API which are not actually associated with views at all, so those can now
be moved to the more appropriate PuglWorld to make this more clear.

The old puglInit() and puglDestroy() functions are preserved for compatibility,
but marked as deprecated.