Windows: Implement size constraints
There are two possible approaches to take here: try to expand dimensions that
are not being explicitly resized (for example expand the bottom when dragging
right), or just stop single-dimension resizes if they would go out of range.

I chose the latter here for two reasons: it's hard to always do something
smooth and unsurprising with the first approach (and it would require more
code), and it can be nice from the user's perspective to easily be able to
resize the window to exactly one of its aspect ratio limits.  For example, it
is very easy to drag pugl_test to 1:1 or 16:9.

In other words, simplicity and user power wins.