Give backends general names
Towards making them opaque and exposing them to the user to decouple the core
library from backends.

The general names mean that it won't be possible to build multiple backends for
one platform into the same binary, but that seems reasonable for now, and it
will make things simpler without needing to add a bunch of dispatch code.  That
will still be possible if it's ever needed, though.