Completely separate backends from platform implementation
This removes PuglContextType and allows the user to pass a backend directly.
Normally this would come from one of the two accessors declared in the headers,
but it would be possible to add backends without changing any of the existing
code at all.

Unfortunately, it is not possible to preserve backwards compatibility and
achieve the ultimate goal of linking only to the required dependencies, so
puglInitContextType() has just been removed.