X11: Support UTF8 in window titles
It's possible that this does bad things if a UTF8 string is used on a system
that does not support it (because XStoreName is still called), but I have no
idea how likely this scenario is.

Leaving it this way because it means the ASCII case will still work everywhere,
and it's easy enough to avoid if this is a problem.