X11: Fix XMoveResizeWindow error check
In the Xlib standard, a return value of 0 is defined as indicating an
error (see /Xlib - C Language X Interface/, Ch. 1, "Errors").
XMoveResizeWindow currently returns 1 on success in conformance with
this. However, the error check in puglSetFrame assumes that 0 indicates
a sucessful resize, causing PUGL_UNKNOWN_ERROR to be returned whenever
the resize succeeds. Flip the check to match the Xlib convention.