Remove immediate dispatch of exposed rects
This was a hack to support only exposing the rects that were explicitly exposed
with puglPostRedisplayRect(), but it caused flaky drawing issues because it
circumvented the deferral of exposure until the end of the loop.  Instead,
simply expand the pending expose to be dispatched later as usual.

This means that only the union will be exposed in the end, so more area might
be drawn than necessary, but this is probably good enough.  If not, we will
have to maintain a set of rects and be more clever about combining them.