Windows: Prevent input in one window from stalling another
This dispatches events on a per-window basic instead of globally, using the
same mark trick as before to bound the number of events dispatched.

After the events are dispatched, all the windows are updated if they have an
invalid region.  This ensures that all windows get drawn every iteration if
necessary, since Windows itself does not send WM_PAINT messages if there is
lots of input activity.