Add create, destroy, map, and unmap events
These can be used to do things when a view is created or destroyed, in
particular set up the GL context in a more controlled way.  Map and unmap
events are also added for when views are shown and hidden so application can
react to this as well.

Towards the deprecation of puglEnterContext() and puglLeaveContext(), which are
prone to abuse.

squash! Remove client event stuff