Unify event loop functions as puglUpdate()
The previous separation between polling and dispatching was a lie, especially
on MacOS where it is impossible to only poll for events without dispatching
anything.  Providing such an API is misleading, and problematic in various
other ways.

So, merge them into a single puglUpdate() function which can do the right thing
on all platforms.  This also adds the behaviour of actually processing all
events in the given time interval, which is almost always what clients actually
want to do when using a positive timeout (naively doing this before caused
terrible input lag).