Separate cached configuration from frame
This was a bad idea and a never-ending source of problems.  The frame
represents what the "current" frame is from a Pugl perspective, but with the
asynchronicity of X11 and other issues this cant be used to filter configure
events.

Instead, simply cache the last configure event that was sent and compare with
that.