Resurrect puglEnterContext() and puglLeaveContext()
Even though loading is now better handled with create events, there are still
situations with OpenGL where it's difficult to avoid explicitly entering and
leaving the context.  So, resurrect these functions, but remove the drawing
parameter to make it clear that they must never be used for drawing.

This breaks the deprecated API since the drawing parameter would be
inappropriate for the stable API, and a decent alternative name isn't
available.