Replace live resize with loop events
This allows the application to control how recursive loops are handled rather
than have Pugl impose behavior which can get in the way.  For example, an
application may not want to continuously animate while being resized, or set up
its rendering differently in this situation.  For example, with Vulkan, setting
up a different swapchain can be necessary for smooth performance while live
resizing on Windows, and Pugl has no ability to do this.

I think it was a mistake to add this timer to Pugl itself, because it was
always a bit of a leaky abstraction, and not very appropriate for a library
that is supposed to be a thin abstraction layer.  Though it almost seemed like
things ran as usual while resizing on Windows and MacOS, the main event loop
being stalled can be confusing, and there was no way to detect this.  This way,
applications must explicitly handle this situation and can implement the
behavior they want without Pugl getting in the way.

This also simplifies the Pugl implementation a bit, which is always nice.