Move puglEnterContext and puglLeaveContext to GL backends
These only do anything for OpenGL, and it seems unlikely that they will ever be
used for anything else.  So, move them to the GL headers to remove clutter from
the core API, and ensure that they are only used in GL applications that
include the appropriate headers and link with a GL backend.

Also add missing C++ bindings.