X11: Fix position of child windows and clean up window creation code
This fixes an issue where the default frame position would be set based on the
screen size for child windows.  This went unnoticed so far because most
plugins, like pugl_embed_demo, explicitly set the frame and so avoided this
code path.

Also generally tidy up puglRealize() along the way to make it a bit more
readable.