Rename event structs in a more readable style
Aside from reading more naturally, this avoids clashes with types that are not
events, like PuglEventFlags.  This is also more consistent with the C++
bindings, where "EventExpose" would be quite strange, for example.

Apologies for the noise.  Aliases to the old names will be preserved in the
deprecated API like other things for a short while.