X11: Fix potential crash in puglGrabFocus()
This is really a mistake in user code, but things shouldn't crash in general.
So, this commit fixes the crash and adds some documentation so that developers
hopefully don't try to grab focus before it makes sense.

The case that was previously a crash will now gracefully fail, that is, the
focus will not be (and can not be) grabbed.