Relax redisplay test to pass on MacOS 11.6
Since upgrading to 11.6 (on an Intel Macbook), this test no longer seems to
pass.  It's unfortunate to not test that small redisplay requests only result
in small exposures, but I don't think the previous strict check reflects
reality.  Exposing more than the application requested, for whatever reason, is
a natural part of window environments, so I don't know if this is something
that can be reliably unit tested.