Fix weird/inconsistent control flow in puglSetSize()
I'm not really sure about directly updating the frame like this - or whether
these fields should even exist in the first place (it's questionable whether
the view is "really" the new size until the configure process is complete) -
but regardless, the same pattern should probably be used for both size and
position.

"Probably", because touching anything like this tends to break some situation
or another somewhere, but such is life.