Windows: Make puglGetScaleFactor always return a reasonable value
This guesses the likely scale factor when the view hasn't been realized (and so
there's no real window to get information for).  It may still be wrong for
multiple display systems, but will always return some scale factor that
"exists" on the system, and should always be correct when there's only one
display.