Separate pugl/glu.h from pugl/gl.h
The GLU library is often shipped in a separate package, so it's possible to
have a system with GL, but no GLU headers.  Since use of GLU isn't
universal (none of the examples use it), its inclusion can break builds that
would otherwise work.  So, move it to a separate wrapper header so this can be
avoided.