X11: Avoid setting PBaseSize hints for top-level windows
By my reading of the spec, pugl's use of this hint was correct.  However, many
window managers break when it's set and use that size as the minimum (even when
an actual minimum is set).  This seems to be a bug based on a misreading of the
spec which has been copy-pasted across many small window manager projects over
the years.

Not exposing the default size is unfortunate, but apparently nobody misses
features based on it and there's nothing we can do about it here, so just deny
the window manager the information entirely.  The hint is still set for
embedded views because (for example) plugin hosts need this information.