MacOS: Fix mouse wheel scroll speed
This just happens to be the factor that results in one tick of a conventional
mouse wheel being a change of 1.0 "lines" as on other platforms (at least for
me, hopefully everywhere).  Moving a wheel quickly does result in quite a lot
of acceleration on MacOS though, so more refinement might be needed here for
consistent event handling across platforms.