Windows: Avoid waiting less than a millisecond
There is API to access the minimum timer resolution, but it seems to always be
1 millisecond, which is also the unit of the MsgWaitForMultipleObjects
parameter, so just avoid the fuss and hardcode this as the minimum.

This should slightly reduce overhead when the main loop is running without much
of a margin until the frame needs to be drawn.