Simplify initial position code and place children at the origin
The code to find the initial size and position for a view had a lot of
duplication, and was inconsistent between platforms.  Flaky positioning and
attempting to center embedded children has historically caused a bunch of
problems in general, particularly since windows can be resized.

So, factor out all the initial size and position code so that the logic is in
one centralized place used by all platforms, and always position embedded
children within their parent at the top/left origin.  For top-level windows,
the view is centered on its transient parent or the screen (as before).  A new
platform function puglGetAncestorCenter() is used to do this in general code.

Also towards a total separation of position and size, towards support for
Wayland, where applications don't dictate their own positions.