Remove PuglRect entirely
This is no longer used internally, and overall, I've found it to be more of a
hassle than a convenience in client code.  It's also arguably questionable to
have a "flat" rectangle when point and area types exist (compare with NSRect,
for example).

Also towards fully separating position from size and discouraging the joining
of them in general, towards Wayland and to discourage common plugin GUI bugs.