Track world state and gracefully handle bad calls
Replaces the dispatchingEvents flag with a more precise world state enum, and
implements checks in some functions.  This isn't implemented universally, but
rather for functions that get abused by unscrupulous developers attempting to
work around bugs, or used in callbacks and event handlers where it's easy to
mistakenly call a function in an unintended context.