Dispatch configure events outside graphics context
This simplifies configure event dispatching, at the cost of breaking code that
was accessing the OpenGL context in the configure handler.  The graphics
context being entered for configure events comes from prehistoric pugl, but
causes a lot of trouble:

- Entering the OpenGL context is expensive on some systems.

- It can't be guaranteed that only a single configure event arrives in a frame.

- It can't be guaranteed that a configure event is the last event to arrive
  before exposure, and other events may influence the graphics context
  configuration (depending on the application).

- Pugl did some internal tricks to avoid or minimize the above problems, but
  this complicated the implementation and could cause inconsistency by
  reordering events from the window system.

- The concept of configuring "in the graphics context" doesn't really make
  sense for other graphics APIs, and implementing it adds unnecessary coupling
  between window management and drawing which can be a performance bottleneck.

So, throw the whole idea out and make configuration a simple event that
is (more or less) forwarded directly from the windowing system to the client.
Clients need to be prepared to receive configure events at any time, including
several between exposures, and OpenGL clients should set up their viewport/etc
immediately before rendering in response to an expose event.