Remove unnecessary call to floor()
I don't think there's any point of using floor() or trunc() here since it's
casting to int anyway.  Coincidentally, this is the last use of libm functions
in the X11 platform implementation, so remove that dependency from the core,
and add it to some examples which were missing it (but got away with it because
it was brought in by the core).