Make Vulkan ProcAddr accessors gracefully handle null
This adds a branch to every implementation, but since failing to create a new
loader is relatively likely (it depends on loading libraries from the system),
I think it's better for these functions to tolerate this.  It's more
resilient (since the caller might still check the returned function pointer and
gracefully handle the failure that way), and more convenient anyway.

These two functions are the whole access API of PuglVulkanLoader, so this also
makes it a consistent rule that accessors that return pointers will gracefully
return null when passed null.