Remove C++ build test
All of this was testing building the library itself as C++ and Objective C++,
which is nice, but I don't think this coverage is worth the significant
maintenance cost, and there's warnings that catch most issues with C++
compatibility.  If necessary, a test on CI that simply passes the sources
directly to the C++ compiler can cover the same thing without all the
complexity.

Towards moving the C++ bindings into a separate project, after which, having
all this C++ build system and warning infrastructure just for this test won't
make much sense (and only make pugl a more annoying dependency/subproject).